Finished Iterim #1 Functionality
User is presented with three hearts ("lives") at the start of each session. If the user fails to sign the presented digit in time, they will lose one coin (like before,) AND lose one heart. Hearts are gained by successfully signing digits (as well as coins) for a maximum of 3 hearts. However, if the user fails to sign 3 digits in a row (losing all hearts,) their coin total resets back to zero.